fix(jans-fido2): Removing multiple Assertion processors #8902
