+++ jans-fido2/server/src/main/java/io/jans/fido2/service/verifier/AssertionVerifier.java
@@ -18,19 +18,27 @@
 
 package io.jans.fido2.service.verifier;
 
+import java.security.PublicKey;
+
+import org.slf4j.Logger;
+
+import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.base.Strings;
+
+import io.jans.fido2.exception.Fido2CompromisedDevice;
+import io.jans.fido2.exception.Fido2RuntimeException;
 import io.jans.fido2.model.assertion.Response;
+import io.jans.fido2.model.auth.AuthData;
+import io.jans.fido2.service.AuthenticatorDataParser;
+import io.jans.fido2.service.Base64Service;
+import io.jans.fido2.service.CoseService;
+import io.jans.fido2.service.DataMapperService;
+import io.jans.fido2.service.util.DigestUtilService;
+import io.jans.fido2.service.util.HexUtilService;
 import io.jans.orm.model.fido2.Fido2AuthenticationData;
 import io.jans.orm.model.fido2.Fido2RegistrationData;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-
-import io.jans.fido2.exception.Fido2RuntimeException;
-import io.jans.fido2.service.processor.assertion.AssertionProcessorFactory;
-import io.jans.fido2.service.processors.AssertionFormatProcessor;
-import org.slf4j.Logger;
-
-import com.fasterxml.jackson.databind.JsonNode;
 
 @ApplicationScoped
 public class AssertionVerifier {
@@ -39,7 +47,31 @@ public class AssertionVerifier {
     private Logger log;
 
     @Inject
-    private AssertionProcessorFactory assertionProcessorFactory;
+    private CoseService coseService;
+
+    @Inject
+    private CommonVerifiers commonVerifiers;
+
+    @Inject
+    private AuthenticatorDataVerifier authenticatorDataVerifier;
+
+    @Inject
+    private UserVerificationVerifier userVerificationVerifier;
+
+    @Inject
+    private AuthenticatorDataParser authenticatorDataParser;
+
+    @Inject
+    private DataMapperService dataMapperService;
+
+    @Inject
+    private Base64Service base64Service;
+
+    @Inject
+    private DigestUtilService digestUtilService;
+
+    @Inject
+    private HexUtilService hexUtilService;
 
     public void verifyAuthenticatorAssertionResponse(Response response, Fido2RegistrationData registration,
                                                      Fido2AuthenticationData authenticationEntity) {
@@ -54,9 +86,38 @@ public class AssertionVerifier {
         String signature = response.getSignature();
 
         log.debug("Authenticator data {}", base64AuthenticatorData);
-        AssertionFormatProcessor assertionProcessor = assertionProcessorFactory.getCommandProcessor(registration.getAttestationType());
+        
 
-        assertionProcessor.process(base64AuthenticatorData, signature, clientDataJson, registration, authenticationEntity);
+        process(base64AuthenticatorData, signature, clientDataJson, registration, authenticationEntity);
     }
+    public void process(String base64AuthenticatorData, String signature, String clientDataJson, Fido2RegistrationData registration,
+            Fido2AuthenticationData authenticationEntity) {
+        AuthData authData = authenticatorDataParser.parseAssertionData(base64AuthenticatorData);
+        commonVerifiers.verifyRpIdHash(authData, registration.getDomain());
 
+        log.debug("User verification option {}", authenticationEntity.getUserVerificationOption());
+        userVerificationVerifier.verifyUserVerificationOption(authenticationEntity.getUserVerificationOption(), authData);
+
+        byte[] clientDataHash = digestUtilService.sha256Digest(base64Service.urlDecode(clientDataJson));
+ 
+        try {
+            int counter = authenticatorDataParser.parseCounter(authData.getCounters());
+            commonVerifiers.verifyCounter(registration.getCounter(), counter);
+            registration.setCounter(counter);
+
+            JsonNode uncompressedECPointNode = dataMapperService.cborReadTree(base64Service.urlDecode(registration.getUncompressedECPoint()));
+            PublicKey publicKey = coseService.createUncompressedPointFromCOSEPublicKey(uncompressedECPointNode);
+
+            log.debug("Uncompressed ECpoint node {}", uncompressedECPointNode);
+            log.debug("EC Public key hex {}", hexUtilService.encodeHexString(publicKey.getEncoded()));
+            log.debug("registration.getSignatureAlgorithm(): "+registration.getSignatureAlgorithm());
+            log.debug("Platform authenticator: "+ registration.getSignatureAlgorithm());
+            authenticatorDataVerifier.verifyAssertionSignature(authData, clientDataHash, signature, publicKey,  registration.getSignatureAlgorithm());
+           
+        } catch (Fido2CompromisedDevice ex) {
+        	throw ex;
+        } catch (Exception ex) {
+            throw new Fido2RuntimeException("Failed to check assertion", ex);
+        }
+    }
 }
